requirements.txt
aiohttp==3.9.1
beautifulsoup4==4.12.2
Flask==3.0.0
Flask-CORS==4.0.0
loguru==0.7.2
python-dotenv==1.0.0
requests==2.31.0
SQLAlchemy==2.0.23 